feat: move notification banner to App.tsx
