+++ frontend/app/chat/components/index.ts
@@ -1,2 +1,1 @@
 export * from "./ChatsList";
-export * from "./NotificationBanner";
