+++ frontend/app/chat/layout.tsx
@@ -6,8 +6,6 @@ import { ChatsProvider } from "@/lib/context/ChatsProvider/chats-provider";
 import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { redirectToLogin } from "@/lib/router/redirectToLogin";
 
-import { NotificationBanner } from "./components";
-
 interface LayoutProps {
   children?: ReactNode;
 }
@@ -23,7 +21,6 @@ const Layout = ({ children }: LayoutProps): JSX.Element => {
     <KnowledgeToFeedProvider>
       <ChatsProvider>
         <ChatProvider>
-          <NotificationBanner />
           <div className="relative h-full w-full flex justify-stretch items-stretch overflow-auto">
             {children}
           </div>
