+++ frontend/lib/components/NotificationBanner/utils.ts
